05_Our_First_Agent_with_LangGraph/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
# Core LangChain and LangGraph
arxiv>=2.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.3.27
langchain-openai>=0.3.27
langgraph>=0.5.1
langsmith>=0.4.4
tavily-python>=0.7.9
# AI and ML
openai>=1.12.0
qdrant-client>=1.7.0
# Data handling
pydantic>=2.5.0
httpx>=0.26.0
python-multipart>=0.0.6
pymupdf>=1.24.0
pillow>=10.0.0
pypdf2>=3.0.0
# Chess libraries
chess>=1.10.0
python-chess>=1.999
# Utilities
loguru>=0.7.0
python-dotenv>=1.0.0
# Development
jupyter>=1.1.1 